Feat/85 hero (#132)
* add hero

* fix styling

* fix css issue on dropdown menu

* Update ProductNavigation.tsx

* Update landingHero.module.css

* add truncate

* fix some styling

* cleaned up old components and fixed some mobile styling

* format megamenu and remove dead code

* Update productNavigation.module.css

* Update index.astro
