+++ src/components/Demos.module.css
@@ -3,6 +3,10 @@
   display: flex;
   flex-direction: column;
   gap: var(--space-6x);
+
+  & > h2 {
+    font-size: 28px;
+  }
 }
 
 .grid {
