+++ src/components/Header/Nav/ProductNavigation/Desktop/MegaMenu.tsx
@@ -1,5 +1,4 @@
 import styles from "./megaMenu.module.css"
-import { useEffect } from "react"
 import ccipLogo from "../../../../../assets/product-logos/ccip-logo.svg"
 import dataFeedsLogo from "../../../../../assets/product-logos/data-feeds-logo.svg"
 import dataStreamsLogo from "../../../../../assets/product-logos/data-streams-logo.svg"
@@ -13,11 +12,7 @@ import nodesLogo from "../../../../../assets/product-logos/nodes-logo.svg"
 import chainlinkLocalLogo from "../../../../../assets/product-logos/chainlink-local-2-logo.svg"
 import creLogo from "../../../../../assets/product-logos/cre-logo.svg"
 import { Typography } from "@chainlink/blocks"
-
-interface MegaMenuProps {
-  cancel: () => void
-  id?: string
-}
+import { clsx } from "../../utils.ts"
 
 const BlueSquare = () => {
   return (
@@ -44,16 +39,10 @@ const GroupItem = ({ data }: { data: GroupItem }) => {
     <a href={data.link} className={styles.groupItem}>
       <img src={data.icon.src} alt={data.title} className={styles.groupItemIcon} />
       <div>
-        <Typography
-          variant="body-semi"
-          style={{
-            fontSize: "18px",
-            fontWeight: 500,
-          }}
-        >
+        <Typography variant="body-semi" className={styles.groupItemTitle}>
           {data.title}
         </Typography>
-        <Typography color="muted" variant="body-s">
+        <Typography color="muted" variant="body-s" className={styles.groupItemDescription}>
           {data.description}
         </Typography>
       </div>
@@ -178,95 +167,78 @@ export const megaMenuSections = {
   },
 }
 
-function MegaMenu({ cancel, id }: MegaMenuProps) {
-  useEffect(() => {
-    const onESC = (ev: KeyboardEvent) => {
-      if (ev.key === "Escape") {
-        cancel()
-      }
-    }
-    window.addEventListener("keyup", onESC, false)
-    return () => {
-      window.removeEventListener("keyup", onESC, false)
-    }
-  }, [])
-
+function MegaMenu({ cancel, isMobile }: { cancel?: () => void; isMobile?: boolean }) {
   return (
-    <div className={styles.megaMenuContainer} id={id}>
-      <div className={styles.wrapper}>
-        <div className={styles.column}>
-          <div className={styles.section}>
-            <header>
-              <BlueSquare />
-              <GroupTitle>{megaMenuSections.orchestration.title}</GroupTitle>
-            </header>
-            <div className={styles.itemList}>
-              {megaMenuSections.orchestration.items.map((link) => (
-                <GroupItem key={link.title} data={link} />
-              ))}
-            </div>
-          </div>
-          <div className={styles.section}>
-            <header>
-              <BlueSquare />
-              <GroupTitle>{megaMenuSections.data.title}</GroupTitle>
-            </header>
-            <div className={styles.itemList}>
-              {megaMenuSections.data.items.map((link) => (
-                <GroupItem key={link.title} data={link} />
-              ))}
-            </div>
-          </div>
+    <div className={clsx(styles.wrapper, isMobile && styles.mobile)} onMouseLeave={cancel}>
+      <div className={styles.row}>
+        <div className={styles.section}>
+          <header>
+            <BlueSquare />
+            <GroupTitle>{megaMenuSections.orchestration.title}</GroupTitle>
+          </header>
+          <li className={styles.itemList}>
+            {megaMenuSections.orchestration.items.map((link) => (
+              <GroupItem key={link.title} data={link} />
+            ))}
+          </li>
         </div>
-
-        <div className={styles.column}>
-          <div className={styles.section}>
-            <header>
-              <BlueSquare />
-              <GroupTitle>{megaMenuSections.interoperability.title}</GroupTitle>
-            </header>
-            <div className={styles.itemList}>
-              {megaMenuSections.interoperability.items.map((link) => (
-                <GroupItem key={link.title} data={link} />
-              ))}
-            </div>
-          </div>
-          <div className={styles.section}>
-            <header>
-              <BlueSquare />
-              <GroupTitle>{megaMenuSections.compute.title}</GroupTitle>
-            </header>
-            <div className={styles.itemList}>
-              {megaMenuSections.compute.items.map((link) => (
-                <GroupItem key={link.title} data={link} />
-              ))}
-            </div>
-          </div>
+        <div className={styles.section}>
+          <header>
+            <BlueSquare />
+            <GroupTitle>{megaMenuSections.interoperability.title}</GroupTitle>
+          </header>
+          <li className={styles.itemList}>
+            {megaMenuSections.interoperability.items.map((link) => (
+              <GroupItem key={link.title} data={link} />
+            ))}
+          </li>
         </div>
+        <div className={styles.section}>
+          <header>
+            <BlueSquare />
+            <GroupTitle>{megaMenuSections.assetManagement.title}</GroupTitle>
+          </header>
+          <li className={styles.itemList}>
+            {megaMenuSections.assetManagement.items.map((link) => (
+              <GroupItem key={link.title} data={link} />
+            ))}
+          </li>
+        </div>
+      </div>
 
-        <div className={styles.column}>
-          <div className={styles.section}>
-            <header>
-              <BlueSquare />
-              <GroupTitle>{megaMenuSections.assetManagement.title}</GroupTitle>
-            </header>
-            <div className={styles.itemList}>
-              {megaMenuSections.assetManagement.items.map((link) => (
-                <GroupItem key={link.title} data={link} />
-              ))}
-            </div>
-          </div>
-          <div className={styles.section}>
-            <header>
-              <BlueSquare />
-              <GroupTitle>{megaMenuSections.other.title}</GroupTitle>
-            </header>
-            <div className={styles.itemList}>
-              {megaMenuSections.other.items.map((link) => (
-                <GroupItem key={link.title} data={link} />
-              ))}
-            </div>
-          </div>
+      <div className={styles.row}>
+        <div className={styles.section}>
+          <header>
+            <BlueSquare />
+            <GroupTitle>{megaMenuSections.data.title}</GroupTitle>
+          </header>
+          <li className={styles.itemList}>
+            {megaMenuSections.data.items.map((link) => (
+              <GroupItem key={link.title} data={link} />
+            ))}
+          </li>
+        </div>
+        <div className={styles.section}>
+          <header>
+            <BlueSquare />
+            <GroupTitle>{megaMenuSections.compute.title}</GroupTitle>
+          </header>
+          <li className={styles.itemList}>
+            {megaMenuSections.compute.items.map((link) => (
+              <GroupItem key={link.title} data={link} />
+            ))}
+          </li>
+        </div>
+        <div className={styles.section}>
+          <header>
+            <BlueSquare />
+            <GroupTitle>{megaMenuSections.other.title}</GroupTitle>
+          </header>
+          <li className={styles.itemList}>
+            {megaMenuSections.other.items.map((link) => (
+              <GroupItem key={link.title} data={link} />
+            ))}
+          </li>
         </div>
       </div>
     </div>
