+++ src/components/Header/Nav/ProductNavigation/Desktop/megaMenu.module.css
@@ -9,30 +9,35 @@
   justify-content: center;
 }
 
+/* applies shadow to dropdown version of mega-menu */
+.megaMenuContainer > .wrapper {
+  box-shadow:
+    4.6px 9.3px 54px 1px rgba(152, 160, 185, 0.04),
+    2.1px 4.3px 32px -0.7px rgba(152, 160, 185, 0.31),
+    1px -1px 3px -0.3px rgba(152, 160, 185, 0.21),
+    0.3px 0.5px 0.8px 0px rgba(152, 160, 185, 0.12);
+}
+
 .wrapper {
   max-width: 1500px;
   width: 100%;
   background-color: white;
   height: fit-content;
   overflow: hidden;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  box-shadow:
-    4.6px 9.3px 54px 1px rgba(152, 160, 185, 0.04),
-    2.1px 4.3px 32px -0.7px rgba(152, 160, 185, 0.31),
-    1px -1px 3px -0.3px rgba(152, 160, 185, 0.21),
-    0.3px 0.5px 0.8px 0px rgba(152, 160, 185, 0.12);
+  display: flex;
+  flex-direction: column;
+
   border: 1px solid var(--border);
 }
 
-.column {
+.row {
   display: grid;
-  grid-template-columns: 1fr;
-  border-right: 1px solid var(--Border-Grid, rgba(22, 37, 65, 0.13));
+  grid-template-columns: repeat(3, 1fr);
+  border-bottom: 1px solid var(--border);
 }
 
-.column:last-child {
-  border-right: none;
+.row:last-child {
+  border-bottom: none;
 }
 
 .wrapper header {
@@ -49,16 +54,21 @@
 .section {
   display: flex;
   flex-direction: column;
+  border-right: 1px solid var(--border);
+}
+
+.section:last-child {
+  border-right: none;
 }
 
 .section header {
-  border-bottom: 1px solid var(--Border-Grid, rgba(22, 37, 65, 0.13));
+  border-bottom: 1px solid var(--border);
 }
 
 .groupItem {
   display: flex;
   align-items: center;
-  padding: var(--space-4x) var(--space-6x);
+  padding: var(--space-6x) var(--space-8x);
   gap: var(--space-4x);
   text-decoration: none;
   transition: background-color 0.2s;
@@ -77,7 +87,59 @@
 .itemList {
   display: flex;
   flex-direction: column;
-  padding: var(--space-2x);
+}
+
+.groupItemDescription {
+  display: -webkit-box;
+  -webkit-line-clamp: 1;
+  line-clamp: 1;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+}
+
+.groupItemTitle {
+  display: -webkit-box;
+  -webkit-line-clamp: 1;
+  line-clamp: 1;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+  font-size: 18px;
+  font-weight: 500;
+}
+
+.wrapper.mobile {
+  border: none;
+  overflow: unset;
+
+  & .row {
+    grid-template-columns: 1fr;
+    border-bottom: none;
+  }
+
+  & .section header {
+    border-bottom: 1px solid var(--border);
+    border-top: 1px solid var(--border);
+  }
+
+  & .section {
+    border-right: none;
+  }
+
+  & .groupItemTitle {
+    display: block;
+    -webkit-line-clamp: unset;
+    line-clamp: unset;
+    -webkit-box-orient: unset;
+    overflow: visible;
+  }
+
+  & .groupItemDescription {
+    display: block;
+    -webkit-line-clamp: unset;
+    line-clamp: unset;
+    -webkit-box-orient: unset;
+    overflow: visible;
+  }
 }
 
 @media screen and (min-width: 768px) {
