+++ src/components/Header/Nav/ProductNavigation/Desktop/productNavigation.module.css
@@ -57,19 +57,6 @@
   opacity: 1;
 }
 
-.megaMenu::after {
-  content: "";
-  position: absolute;
-  bottom: -35px;
-  right: 35%;
-  transform: translateX(-50%) rotate(45deg);
-  width: 20px;
-  height: 20px;
-  z-index: 100;
-  background-color: white;
-  padding: var(--space-2x);
-}
-
 @media screen and (min-width: 992px) {
   .root {
     display: flex;
