--- src/components/Header/Nav/ProductNavigation/Mobile/megaMenu.module.css
@@ -1,142 +1,0 @@
-.wrapper {
-  width: 100%;
-  border-radius: 0.5rem;
-  display: flex;
-  flex-direction: column;
-  max-height: calc(100vh - 68px);
-}
-
-.wrapper h2 {
-  padding-bottom: var(--space-3x);
-  color: var(--gray-500);
-  text-transform: uppercase;
-
-  border-bottom: 1px solid var(--gray-200);
-}
-
-.wrapper h3 {
-  font-family: var(--font-family-text);
-  color: var(--gray-900);
-}
-
-.wrapper p {
-  color: var(--gray-500);
-  margin-top: var(--space-2x);
-  margin-bottom: var(--space-3x);
-  line-height: var(--space-6x);
-}
-
-.resourcesMenuContentMain {
-  padding: var(--space-6x) var(--space-8x) var(--space-4x) var(--space-8x);
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-6x);
-}
-
-.resourcesMenuContentRow {
-  display: flex;
-  flex-direction: column;
-}
-
-.verticalDivider {
-  border-left: 1px solid var(--gray-300);
-  margin: 0 var(--space-3x);
-}
-
-.links {
-  margin-bottom: var(--space-6x);
-  margin-left: var(--space-8x);
-}
-
-.links a {
-  font-weight: var(--font-weight-medium);
-  color: var(--color-text-link);
-}
-
-.resourcesMenuContentFeatured {
-  padding-top: var(--space-16x);
-  padding-left: var(--space-8x);
-  padding-bottom: var(--space-16x);
-  padding-right: var(--space-16x);
-  background-color: var(--gray-100);
-}
-
-.megaMenuButton {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: var(--space-4x) var(--space-1x);
-  cursor: pointer;
-}
-
-.megaMenuButton span {
-  font-size: var(--space-6x);
-  font-weight: var(--font-weight-medium);
-  color: var(--gray-500);
-}
-
-.megaMenuLink {
-  display: flex;
-  align-items: center;
-  gap: var(--space-2x);
-}
-
-.megaMenuLink h3 {
-  margin-bottom: 0;
-}
-
-.megaMenuLink img {
-  width: var(--space-5x);
-  height: var(--space-5x);
-}
-
-.featuredImage {
-  width: 100%;
-}
-
-.divider {
-  border-bottom: 1px solid var(--gray-200);
-  margin: var(--space-8x) auto;
-}
-
-.bottomLinks {
-  padding: var(--space-6x) 0;
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-8x);
-  border-top: 1px solid var(--gray-200);
-}
-
-.bottomLinks div {
-  display: flex;
-  align-items: center;
-  gap: var(--space-2x);
-  line-height: var(--space-6x);
-}
-
-.bottomLinks a {
-  color: var(--gray-900);
-}
-
-/* TODO: change with a var when available in the chainlink repo */
-.bottomLink a:hover {
-  color: #0847f7;
-}
-
-@media screen and (min-width: 768px) {
-  .megaMenuContainer {
-    display: flex;
-  }
-}
-
-@media screen and (min-width: 1280px) {
-  .wrapper::before {
-    left: 31%;
-  }
-}
-
-@media screen and (min-width: 1565px) {
-  .wrapper::before {
-    left: 35%;
-  }
-}
