+++ src/components/Header/Nav/ProductNavigation/Mobile/productNavigation.module.css
@@ -34,6 +34,7 @@
   width: 200vw;
   display: flex;
   transition: transform 0.3s;
+  height: 100%;
 }
 
 .main {
@@ -71,11 +72,11 @@
   overflow: scroll;
   list-style: none;
   width: 100vw;
-  padding-bottom: 56px;
 }
 .subProductContent {
   display: flex;
   flex-direction: column;
+  height: 100%;
 }
 
 .subProductContentTitle {
