+++ src/components/JourneyCards/JourneyCards.astro
@@ -198,7 +198,7 @@ const tabs = columns.map((column) => ({
   }
 
   .section-title {
-    font-size: 32px;
+    font-size: 28px;
     margin-bottom: var(--space-10x);
   }
 
@@ -247,6 +247,12 @@ const tabs = columns.map((column) => ({
     .journey-card h4 {
       font-size: 16px;
       line-height: 22px;
+    }
+  }
+
+  @media screen and (min-width: 62em) {
+    .section-title {
+      font-size: 32px;
     }
   }
 </style>
