+++ src/components/LandingHero/LandingHero.astro
@@ -1,0 +1,21 @@
+---
+import { Typography } from "@chainlink/blocks"
+import MegaMenu from "../Header/Nav/ProductNavigation/Desktop/MegaMenu"
+import styles from "./landingHero.module.css"
+---
+
+<section class={styles.wrapper}>
+  <div class={styles.content}>
+    <Typography variant="h1" className={styles.title}>Explore docs & resources</Typography>
+    <section class={styles.menu}>
+      <MegaMenu client:only="react" />
+    </section>
+  </div>
+
+  <div class={styles.backgrounds}>
+    <div class={styles.heroLeft}><img src="/images/hero-dotted.png" alt="" class={styles.heroDotted} /></div>
+    <div class={styles.heroRight}>
+      <img src="/images/hero-1.png" alt="" class={styles.heroImage} />
+    </div>
+  </div>
+</section>
