+++ src/components/LandingHero/landingHero.module.css
@@ -1,0 +1,69 @@
+.content {
+  padding: var(--space-16x);
+  max-width: 1440px;
+  width: 100%;
+  margin: 0 auto;
+  position: absolute;
+  inset: 0;
+  z-index: 5;
+}
+
+.backgrounds {
+  display: flex;
+  height: 100%;
+}
+
+.title {
+  font-weight: 400;
+  margin-bottom: var(--space-8x);
+}
+
+.wrapper {
+  background-color: var(--muted);
+  position: relative;
+  inset: 0;
+  height: 746px;
+}
+
+.heroImage {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.heroDotted {
+  width: 100%;
+  object-fit: cover;
+}
+.heroLeft {
+  display: flex;
+  align-items: end;
+}
+
+.heroRight,
+.heroLeft {
+  width: 50%;
+  height: 100%;
+}
+
+@media screen and (max-width: 991px) {
+  .wrapper {
+    background-color: transparent;
+    height: auto;
+    padding: 36px var(--space-10x);
+  }
+  .backgrounds,
+  .content .menu {
+    display: none;
+  }
+
+  .content {
+    position: unset;
+    padding: 0;
+  }
+
+  .title {
+    font-size: 40px;
+    margin-bottom: 0;
+  }
+}
