--- src/features/landing/sections/JourneyCards.astro
@@ -1,211 +1,0 @@
----
-import { Tag, Typography } from "@chainlink/blocks"
-
-const columns = [
-  {
-    title: "Learn & Explore",
-    items: [
-      {
-        title: "Explore Cross-Chain Interoperability with CCIP",
-        description: "Learn cross-chain concepts, workflows, and real-world use cases.",
-        badge: "CCIP",
-        href: "/",
-      },
-      {
-        title: "Understand How Data Feeds Power dApps",
-        description: "See how oracle data feeds deliver price feeds and reference data.",
-        badge: "DATA FEEDS",
-        href: "/",
-      },
-      {
-        title: "Learn How Data Streams Deliver Real-Time Data",
-        description: "Understand how low-latency streams support time-sensitive applications.",
-        badge: "CCIP",
-        href: "/",
-      },
-      {
-        title: "Discover Off-Chain Compute with Functions",
-        description: "Learn how Functions connect smart contracts to APIs and custom logic.",
-        badge: "DATA FEEDS",
-        href: "/",
-      },
-    ],
-  },
-  {
-    title: "Build & Integrate",
-    items: [
-      {
-        title: "Build Cross-Chain Apps with CCIP Tutorials",
-        description: "Follow step-by-step guides with language switching (EVM, Rust, Move, etc.).",
-        badge: "CCIP",
-        href: "/",
-      },
-      {
-        title: "Integrate Data Feeds into Smart Contracts",
-        description: "Plug feeds into your apps with examples and addresses.",
-        badge: "DATA FEEDS",
-        href: "/",
-      },
-      {
-        title: "Implement Real-Time Use Cases with Data Streams",
-        description: "Use low-latency data in trading, gaming, and other live applications.",
-        badge: "CCIP",
-        href: "/",
-      },
-      {
-        title: "Connect Contracts to APIs with Functions",
-        description: "Add external data and custom logic to your dApps.",
-        badge: "DATA FEEDS",
-        href: "/",
-      },
-    ],
-  },
-  {
-    title: "Operate & Scale",
-    items: [
-      {
-        title: "Monitor CCIP Transactions in Real Time",
-        description: "Track the progress and status of cross-chain transactions.",
-        badge: "CCIP",
-        href: "/",
-      },
-      {
-        title: "Stay Up to Date with Data Feeds",
-        description: "Rely on changelogs and schema updates for accuracy.",
-        badge: "DATA FEEDS",
-        href: "/",
-      },
-      {
-        title: "Deliver Reliable Low-Latency Data with Streams",
-        description: "Operate Data Streams at scale for critical, time-sensitive use cases.",
-        badge: "CCIP",
-        href: "/",
-      },
-      {
-        title: "Scale and Optimize Functions",
-        description: "Debug, manage workloads, and grow your applications.",
-        badge: "DATA FEEDS",
-        href: "/",
-      },
-    ],
-  },
-]
----
-
-<div class="journey-cards">
-  {
-    columns.map((column) => (
-      <div class="journey-column">
-        <header class="column-header">
-          <Typography variant="h5" className="column-title">
-            {column.title}
-          </Typography>
-        </header>
-        {column.items.map((item) => (
-          <a href={item.href} class="journey-card">
-            <div class="card-content">
-              <Typography variant="body-semi">{item.title}</Typography>
-              <Typography variant="body-s" color="muted">
-                {item.description}
-              </Typography>
-            </div>
-
-            <footer class="journey-footer">
-              <Tag size="sm" className="footer-tag">
-                <Typography variant="code-s">{item.badge}</Typography>
-              </Tag>
-              <img src="/assets/icons/upper-right-arrow.svg" class="footer-icon" />
-            </footer>
-          </a>
-        ))}
-      </div>
-    ))
-  }
-</div>
-
-<style>
-  .journey-cards {
-    display: grid;
-    grid-template-columns: 1fr;
-  }
-
-  .journey-column {
-    display: flex;
-    flex-direction: column;
-    border-left: 1px solid var(--border);
-  }
-
-  .journey-card {
-    gap: var(--space-6x);
-    padding: var(--space-6x);
-  }
-
-  .journey-card:hover {
-    background-color: var(--muted);
-
-    .footer-tag {
-      background-color: var(--background) !important;
-    }
-
-    .footer-icon {
-      opacity: 1;
-    }
-  }
-
-  .card-content {
-    display: flex;
-    flex-direction: column;
-    gap: var(--space-2x);
-    margin-bottom: var(--space-8x);
-  }
-
-  .journey-footer {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-  }
-
-  .footer {
-    padding-top: var(--space-1x);
-  }
-  .footer-icon {
-    height: 12px;
-    width: 12px;
-    opacity: 0;
-  }
-
-  .card-badge {
-    display: inline-flex;
-    align-items: center;
-    padding: var(--space-1x) var(--space-2x);
-    background-color: var(--gray-100);
-    border-radius: 4px;
-    font-size: 11px;
-    font-weight: 500;
-    text-transform: uppercase;
-    color: var(--gray-600);
-    letter-spacing: 0.5px;
-    align-self: flex-start;
-  }
-
-  .column-header {
-    padding: var(--space-2x) var(--space-6x);
-    border-left: 3px solid var(--brand);
-  }
-
-  @media (min-width: 50em) {
-    .journey-cards {
-      grid-template-columns: repeat(3, 1fr);
-    }
-
-    .column-title {
-      font-size: 22px;
-      line-height: 26px;
-    }
-
-    .journey-card h4 {
-      font-size: 16px;
-      line-height: 22px;
-    }
-  }
-</style>
