--- src/features/landing/sections/ProductTabs.astro
@@ -1,8 +1,0 @@
----
-import { Tabs } from "../components/Tabs"
-import productTabs from "./ProductTabs.module.css"
----
-
-<div class={productTabs.container}>
-  <Tabs client:idle />
-</div>
