--- src/features/landing/sections/ProductTabs.module.css
@@ -1,4 +1,0 @@
-.container {
-  padding-top: var(--space-6x);
-  padding-bottom: 60px;
-}
